# REGR: perf regression in Series.combine_first